requirements.txt
dash
dash-bootstrap-components
flake8
jsonschema
pyyaml